requirements.txt
langchain==0.2.16
langchain-community==0.2.17
fastembed==0.3.6 
langchain-groq==0.1.10
crewai==0.63.6
python-dotenv
pygame==2.6.0